requirements.txt
streamlit-pydantic
pydantic==1.10.14